fix: 滚轮跨屏、A 键与 Ctrl+A、Windows 锁屏注入问题
- capture(macOS): 触控板像素增量转换为行单位、滚轮单格事件聚类、惯性阻尼，修复跨屏滚动异常
- capture(Windows): 使用 LLMHF_INJECTED(0x01) 过滤自身注入的鼠标事件，避免回环转发
- inject(Windows): SendInput 滚轮增量不再左移入高字，修复 65536 倍放大导致整页滚动
- keymap: 用 0xFFFF 作无键码哨兵，修复 macOS 虚拟键码 0x00('A') 被当作无效值导致 A/Cmd+A 失效
- winsec: 精简锁屏代理诊断日志

+++ platform/inject/inject_windows.go
@@ -155,9 +155,13 @@ func (w *windowsInjector) Button(b event.MouseButton, down bool) error {
 	return sendInput(inp)
 }
 
-// wheelData encodes a signed wheel delta in the high WORD of mouseData.
+// wheelData encodes the wheel delta for SendInput. MOUSEINPUT.mouseData
+// carries the full signed 32-bit amount (120 per notch) — unlike the
+// low-level hook / WM_MOUSEWHEEL path where the delta lives in the HIGH
+// word. Shifting the value here (as one must when decoding a hook event)
+// multiplied the delta by 65536, making one notch scroll entire pages.
 func wheelData(lines int32) uint32 {
-	return uint32(uint16(int16(lines * wheelDelta)))
+	return uint32(int32(lines * wheelDelta))
 }
 
 func (w *windowsInjector) Scroll(dx, dy int32) error {
